+++ packages/pi/src/loader/config.ts
@@ -8,7 +8,6 @@ export {
 	VERSION,
 } from "#pi/loader/app";
 export {
-	getChangelogPath,
 	getDocsPath,
 	getPackageDir,
 	getReadmePath,